requirements.txt
pyyaml
flask
pytest

# AI Model Integrations
openai>=1.0.0
anthropic>=0.7.0
requests>=2.31.0
aiohttp>=3.9.0
tenacity>=8.2.0

# Database
sqlalchemy>=2.0.0

# Caching
redis>=5.0.0
hiredis>=2.2.0

# Monitoring
prometheus-client>=0.19.0

# Authentication
pyjwt>=2.8.0
passlib>=1.7.4
python-multipart>=0.0.6

# Security & Authentication
bcrypt>=4.0.0
python-dotenv>=1.0.0

# Data Science & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ML Datasets
datasets>=2.14.0

# Vector Storage (optional - FAISS for development)
# faiss-cpu>=1.7.4  # Uncomment for local vector storage

# WebSocket Support (NOT YET IMPLEMENTED - code exists but not integrated)
# flask-socketio>=5.3.0
# python-socketio>=5.10.0
# eventlet>=0.33.0
